+++ .gitignore
@@ -131,3 +131,4 @@ dmypy.json
 # Personal.
 .vscode/
 **/*.mp3
+.DS_store
